+++ cdk/lib/stripe-disputes.ts
@@ -90,7 +90,7 @@ export class StripeDisputes extends GuStack {
 			functionName: nameWithStageProducer,
 			loggingFormat: LoggingFormat.TEXT,
 			fileName: `${app}.zip`,
-			handler: 'index.handler',
+			handler: 'producer.handler',
 			runtime: nodeVersion,
 			memorySize: 1024,
 			timeout: Duration.seconds(300),
@@ -124,7 +124,7 @@ export class StripeDisputes extends GuStack {
 				functionName: nameWithStageConsumer,
 				loggingFormat: LoggingFormat.TEXT,
 				fileName: `${app}.zip`,
-				handler: 'index.handler',
+				handler: 'consumer.handler',
 				runtime: nodeVersion,
 				memorySize: 1024,
 				timeout: Duration.seconds(300),
