+++ handlers/stripe-disputes/package.json
@@ -4,7 +4,7 @@
     "test": "jest --group=-integration",
     "it-test": "jest --group=integration",
     "type-check": "tsc --noEmit",
-    "build": "esbuild --bundle --platform=node --target=node18 --outfile=target/index.js src/index.ts",
+    "build": "mkdir -p target && esbuild --bundle --platform=node --target=node18 --outfile=target/producer.js src/producer.ts && esbuild --bundle --platform=node --target=node18 --outfile=target/consumer.js src/consumer.ts",
     "lint": "eslint src/**/*.ts test/**/*.ts",
     "package": "pnpm type-check && pnpm lint && pnpm check-formatting && pnpm test && pnpm build && cd target && zip -qr stripe-disputes.zip ./*.js",
     "check-formatting": "prettier --check **.ts",
