--- handlers/stripe-disputes/src/handlers/index.ts
@@ -1,2 +1,0 @@
-export * from './disputeCreatedHandler';
-export * from './disputeClosedHandler';
